requirements.txt
Django==3.1.5
django-allauth==0.44.0
django-crispy-forms==1.10.0
django-environ==0.4.5
django-bootstrap4==2.3.1
django-bootstrap-datepicker-plus==3.0.5
django-widget-tweaks==1.4.8
django-braces==1.14.0
whitenoise==5.2.0
gunicorn==20.0.4
psycopg2==2.8.5